refactor: move monument domain to separate module and update imports and dependencies
