+++ content/src/main/java/pt/estga/content/controllers/MarkOccurrenceController.java
@@ -4,16 +4,13 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import pt.estga.content.dtos.*;
-import pt.estga.content.mappers.MarkMapper;
 import pt.estga.content.mappers.MarkOccurrenceMapper;
-import pt.estga.content.mappers.MonumentMapper;
 import pt.estga.content.services.MarkOccurrenceQueryService;
 import pt.estga.content.services.MarkSearchService;
 import pt.estga.detection.model.DetectionResult;
