+++ decision/src/test/java/pt/estga/decision/listeners/SubmissionEventListenerTest.java
@@ -7,11 +7,11 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import pt.estga.content.entities.Mark;
 import pt.estga.content.entities.MarkOccurrence;
-import pt.estga.content.entities.Monument;
 import pt.estga.content.services.MarkOccurrenceService;
-import pt.estga.content.services.MonumentService;
 import pt.estga.decision.services.DecisionServiceFactory;
 import pt.estga.file.entities.MediaFile;
+import pt.estga.monument.Monument;
+import pt.estga.monument.MonumentService;
 import pt.estga.submission.entities.MarkOccurrenceSubmission;
 import pt.estga.submission.events.SubmissionAcceptedEvent;
 import pt.estga.submission.repositories.MarkOccurrenceSubmissionRepository;
