--- monument/src/main/java/pt/estga/monument/MonumentAutoConfiguration.java
@@ -1,4 +1,0 @@
-package pt.estga.monument;
-
-public class MonumentAutoConfiguration {
-}
