+++ monument/src/main/java/pt/estga/monument/MonumentRepository.java
@@ -1,4 +1,4 @@
-package pt.estga.content.repositories;
+package pt.estga.monument;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -7,7 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import pt.estga.content.entities.Monument;
 
 import java.util.Optional;
