+++ monument/src/main/java/pt/estga/monument/MonumentService.java
@@ -1,16 +1,9 @@
-package pt.estga.content.services;
+package pt.estga.monument;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import pt.estga.content.entities.Monument;
-import pt.estga.content.repositories.MonumentRepository;
-import pt.estga.territory.entities.AdministrativeDivision;
-import pt.estga.territory.services.AdministrativeDivisionQueryService;
-import pt.estga.territory.services.AdministrativeDivisionService;
 
-import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 
 @Service
