+++ monument/src/main/java/pt/estga/monument/dots/MonumentDto.java
@@ -1,4 +1,4 @@
-package pt.estga.content.dtos;
+package pt.estga.monument.dots;
 
 import pt.estga.territory.dtos.AdministrativeDivisionDto;
