+++ monument/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-pt.estga.monument.MonumentAutoConfiguration
+pt.estga.monument.config.MonumentAutoConfiguration
