+++ submission/src/main/java/pt/estga/submission/dtos/ProposalWithRelationsDto.java
@@ -1,8 +1,8 @@
 package pt.estga.submission.dtos;
 
 import pt.estga.content.dtos.MarkDto;
-import pt.estga.content.dtos.MonumentDto;
 import pt.estga.file.dtos.MediaFileDto;
+import pt.estga.monument.dots.MonumentDto;
 import pt.estga.submission.enums.SubmissionStatus;
 import pt.estga.submission.enums.SubmissionSource;
 import pt.estga.user.dtos.UserDto;
