+++ submission/src/main/java/pt/estga/submission/entities/MarkOccurrenceSubmission.java
@@ -4,8 +4,8 @@ import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.Type;
 import pt.estga.content.entities.Mark;
-import pt.estga.content.entities.Monument;
 import pt.estga.file.entities.MediaFile;
+import pt.estga.monument.Monument;
 import pt.estga.shared.utils.PgVectorType;
 import pt.estga.submission.enums.SubmissionSource;
 import pt.estga.submission.enums.SubmissionStatus;
