+++ submission/src/main/java/pt/estga/submission/services/MarkOccurrenceSubmissionSubmitService.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import pt.estga.content.entities.Mark;
-import pt.estga.content.entities.Monument;
 import pt.estga.file.entities.MediaFile;
 import pt.estga.file.services.MediaService;
+import pt.estga.monument.Monument;
 import pt.estga.submission.dtos.MarkOccurrenceProposalCreateDto;
 import pt.estga.submission.entities.MarkOccurrenceSubmission;
 import pt.estga.submission.enums.SubmissionSource;
