+++ submission/src/test/java/pt/estga/submission/services/SubmissionScoringServiceTest.java
@@ -7,8 +7,8 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.mockito.junit.jupiter.MockitoSettings;
 import org.mockito.quality.Strictness;
-import pt.estga.content.entities.Monument;
 import pt.estga.file.entities.MediaFile;
+import pt.estga.monument.Monument;
 import pt.estga.submission.config.SubmissionDecisionProperties;
 import pt.estga.submission.entities.MarkOccurrenceSubmission;
 import pt.estga.submission.repositories.MarkOccurrenceSubmissionRepository;
